pyproject.toml
[project]
name = "planwijs"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "alembic>=1.18.4",
    "asyncpg>=0.31.0",
    "cuid2>=2.0.1",
    "fastapi[standard]>=0.135.1",
    "greenlet>=3.3.2",
    "httpx>=0.28.1",
    "pydantic-ai>=1.70.0",
    "pydantic-settings>=2.13.1",
    "python-jose[cryptography]>=3.5.0",
    "sqlmodel>=0.0.37",
    "beautifulsoup4>=4.12.0",
    "lxml>=5.0.0",
    "alembic-postgresql-enum>=1.10.0",
    "boto3>=1.42.73",
    "structlog>=24.4.0",
]

[dependency-groups]
dev = [
    "ty>=0.0.24",
]

[tool.ty.src]
exclude = ["alembic/"]
